// src/app/interfaces/object.ts
export interface Object {
  class?: string;
  className?: any;
  createDate?: Date;
  environmentId?: number;
  identifier?: string;
  imageUrl?: string;
  meta?: any;
  modifierAccount?: any;
  modifyDate?: Date;
  name?: string;
  id?: number;
  objects?: any;
  parentIdentifierObjectId?: number;
  parentObject?: any;
  parentObjectId?: number;
  state?: string;
  subclass?: string;
}